backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
psycopg2-binary
flask-sqlalchemy
python-dotenv
pytest==8.1.1